Fixed phone number not showing on profile, and added formatting
